refactor: convert TokenMetadata to a type union
