+++ packages/shared/lib/core/network/constants/base-token.constant.ts
@@ -1,7 +1,7 @@
-import { TokenMetadata } from '@core/wallet'
+import { IBaseToken } from '@core/wallet'
 import { NetworkProtocol } from '../enums'
 
-export const BASE_TOKEN: Readonly<{ [key in NetworkProtocol]: TokenMetadata }> = {
+export const BASE_TOKEN: Readonly<{ [key in NetworkProtocol]: IBaseToken }> = {
     [NetworkProtocol.IOTA]: {
         name: 'IOTA',
         tickerSymbol: 'MIOTA',
@@ -9,7 +9,6 @@ export const BASE_TOKEN: Readonly<{ [key in NetworkProtocol]: TokenMetadata }> =
         decimals: 0,
         subunit: null,
         useMetricPrefix: true,
-        primaryColor: '#6E82A4',
     },
     [NetworkProtocol.Shimmer]: {
         name: 'Shimmer',
@@ -18,7 +17,5 @@ export const BASE_TOKEN: Readonly<{ [key in NetworkProtocol]: TokenMetadata }> =
         decimals: 6,
         subunit: 'glow',
         useMetricPrefix: false,
-        primaryColor: '#25DFCA',
-        url: 'https://shimmer.network/',
     },
 }
