+++ packages/shared/lib/core/network/interfaces/network.interface.ts
@@ -1,4 +1,4 @@
-import { TokenMetadata } from '@core/wallet'
+import { IBaseToken } from '@core/wallet'
 import { NetworkId } from '../enums'
 import { IProtocol } from './protocol.interface'
 
@@ -11,5 +11,5 @@ export interface INetwork {
     id: NetworkId
     name: string
     protocol: IProtocol
-    baseToken: TokenMetadata
+    baseToken: IBaseToken
 }
