+++ packages/shared/lib/core/profile/actions/active-profile/getBaseToken.ts
@@ -1,9 +1,9 @@
 import { BASE_TOKEN } from '@core/network/constants'
 import { activeProfile } from '@core/profile/stores'
-import { TokenMetadata } from '@core/wallet'
+import { IBaseToken } from '@core/wallet'
 import { get } from 'svelte/store'
 
-export function getBaseToken(): TokenMetadata {
+export function getBaseToken(): IBaseToken {
     const $activeProfile = get(activeProfile)
     return BASE_TOKEN?.[$activeProfile?.networkProtocol]
 }
