+++ packages/shared/lib/core/wallet/interfaces/index.ts
@@ -17,4 +17,3 @@ export * from './mint-token-details.interface'
 export * from './wrapped-output.interface'
 export * from './persisted-asset.interface'
 export * from './persisted-assets.interface'
-export * from './token-metadata.interface'
