+++ packages/shared/lib/core/wallet/interfaces/persisted-asset.interface.ts
@@ -1,5 +1,5 @@
 import { AssetVerification } from '../types'
-import { TokenMetadata } from './token-metadata.interface'
+import { TokenMetadata } from '../types/token-metadata.type'
 
 export interface IPersistedAsset {
     id: string
