--- packages/shared/lib/core/wallet/interfaces/token-metadata.interface.ts
@@ -1,14 +1,0 @@
-export interface TokenMetadata {
-    name: string
-    unit: string
-    decimals: number
-    description?: string
-    subunit?: string | null
-    tickerSymbol?: string
-    useMetricPrefix?: boolean
-    url?: string
-    logo?: string
-    logoUrl?: string
-    primaryColor?: string
-    secondaryColor?: string
-}
