+++ packages/shared/lib/core/wallet/types/index.ts
@@ -3,6 +3,6 @@ export * from './asset-verification.type'
 export * from './output.type'
 export * from './sender-info.type'
 export * from './subject.type'
-
 export * from './activities'
 export * from './new-transaction-details.type'
+export * from './token-metadata.type'
