+++ packages/shared/lib/core/wallet/types/token-metadata.type.ts
@@ -1,0 +1,5 @@
+import { TokenStandard } from '../enums'
+import { IBaseToken } from '../interfaces/base-token.interface'
+import { IIrc30Metadata } from '../interfaces/irc30-metadata.interface'
+
+export type TokenMetadata = (IBaseToken & { standard: TokenStandard.BaseCoin }) | IIrc30Metadata
