+++ packages/shared/lib/core/wallet/utils/convertToRawAmount.ts
@@ -3,10 +3,11 @@ import Big from 'big.js'
 import { parseCurrency } from '@core/i18n'
 import { IOTA_UNIT_MAP } from '@core/utils'
 
-import { TokenMetadata } from '../interfaces'
+import { TokenMetadata } from '../types'
 import { MAX_SUPPORTED_DECIMALS } from '../constants/max-supported-decimals.constants'
+import { TokenStandard } from '../enums'
 
-export function convertToRawAmount(amount: string, tokenMetadata: TokenMetadata, unit = tokenMetadata?.unit): Big {
+export function convertToRawAmount(amount: string, tokenMetadata: TokenMetadata, unit?: string): Big {
     if (amount) {
         const parsedAmount = parseCurrency(amount)
         return convertToRawAmountFromMetadata(parsedAmount, tokenMetadata, unit)
@@ -16,13 +17,14 @@ export function convertToRawAmount(amount: string, tokenMetadata: TokenMetadata,
 }
 
 function convertToRawAmountFromMetadata(amount: number, tokenMetadata: TokenMetadata, unit: string): Big {
-    if (tokenMetadata.useMetricPrefix) {
+    const isBaseToken = tokenMetadata?.standard === TokenStandard.BaseCoin
+    if (isBaseToken && tokenMetadata.useMetricPrefix) {
         return Big(amount * IOTA_UNIT_MAP?.[unit?.substring(0, 1)] ?? 0)
     } else {
-        if (unit && unit === tokenMetadata.unit) {
+        if (unit && isBaseToken && unit === tokenMetadata.unit) {
             const decimals = Math.min(tokenMetadata.decimals, MAX_SUPPORTED_DECIMALS)
             return Big(amount).mul(Big(10).pow(decimals))
-        } else if (unit === tokenMetadata.subunit) {
+        } else if (isBaseToken && unit === tokenMetadata.subunit) {
             return Big(amount)
         }
     }
