+++ packages/shared/lib/core/wallet/utils/formatTokenAmountBestMatch.ts
@@ -1,8 +1,9 @@
 import { formatNumber } from '@core/i18n'
 import { getIotaUnit, IOTA_UNIT_MAP } from '@core/utils'
 
-import { TokenMetadata } from '../interfaces'
+import { TokenMetadata } from '../types'
 import { formatTokenAmountDefault } from './formatTokenAmountDefault'
+import { TokenStandard } from '../enums'
 
 export function formatTokenAmountBestMatch(
     amount: number,
@@ -10,10 +11,12 @@ export function formatTokenAmountBestMatch(
     overrideDecimalPlaces?: number,
     withUnit = true
 ): string {
+    const isBaseToken = tokenMetadata?.standard === TokenStandard.BaseCoin
+
     let amountWithoutUnit: string
     let amountWithUnit: string
 
-    if (tokenMetadata?.useMetricPrefix) {
+    if (isBaseToken && tokenMetadata?.useMetricPrefix) {
         const metricUnit = getIotaUnit(amount)
         const maxDecimals = overrideDecimalPlaces ?? IOTA_UNIT_MAP[metricUnit].decimalPlaces
         const convertedAmount = amount / IOTA_UNIT_MAP[metricUnit].value
@@ -21,7 +24,7 @@ export function formatTokenAmountBestMatch(
         amountWithUnit = amountWithoutUnit + ' ' + metricUnit + tokenMetadata.unit
     } else {
         amountWithoutUnit = !isNaN(amount) ? formatTokenAmountDefault(amount, tokenMetadata) : '0'
-        amountWithUnit = amountWithoutUnit + (tokenMetadata?.unit ? ' ' + tokenMetadata.unit : '')
+        amountWithUnit = amountWithoutUnit + (isBaseToken && tokenMetadata?.unit ? ' ' + tokenMetadata.unit : '')
     }
 
     if (withUnit) {
